Add int casts in std::pow to keep g++ 4.1.2 happy.
git-svn-id: svn+ssh://software.sandia.gov/svn/pecos/trunk@289 ccfab4c5-ad45-0410-a689-f63b6aebdeca

+++ src/InterpPolyApproximation.cpp
@@ -1067,7 +1067,7 @@ void InterpPolyApproximation::compute_global_sensitivity()
 void InterpPolyApproximation::get_subsets()
 {
   // includes the "zero" set
-  int num_subsets = (int)std::pow(2.0,numVars);
+  int num_subsets = (int)std::pow(2.,(int)numVars);
 
   for (int i=1; i<num_subsets; i++) {
     // find all constituent subsets of set i and store
